clean up the backend interfaces and testing
