+++ README.md
@@ -2,6 +2,8 @@
 
 CBSafeKit is a lightweight helper package, to help Swift programmers safely instantiate various CoreBluetooth objects.
 
+_"Disclaimer: You can objectively see I am no expert with Objective-C"_
+
 ## The problem
 
 Instantiating some CoreBluetooth objects can cause your program to raise an NSInternalInconsistencyException, which cannot be handled in a Swift do-catch. Unfortunately, there is very limited documentation for what is/isn't valid input for these types. As such, you might be tempted to write a lot of code to attempt to validate objects that _could_ be valid.
@@ -45,9 +47,11 @@ func doSomething() {
 * Returns `CBMutableService?`
 * Does not throw
 
+### SafeCBMutableCharacteristic()
+
 ## Documented Dangers
 
-### [CBUUID.init(string: String)](https://developer.apple.com/documentation/corebluetooth/cbuuid/1519025-init)
+### [CBUUID.init(string:)](https://developer.apple.com/documentation/corebluetooth/cbuuid/1519025-init)
 
 The following inputs will cause this method to throw a `NSInternalInconsistencyException`:
 * If `string` length is 4 or 8:
@@ -57,13 +61,12 @@ The following inputs will cause this method to throw a `NSInternalInconsistencyE
     * All other characters are not hexadecimal (case insensitive)
 * Else, the error will be thrown
 
-### [CBMutableService.init(type: CBUUID, primary: Bool)](https://developer.apple.com/documentation/corebluetooth/cbmutableservice/1434330-init)
+### [CBMutableService.init(type:primary:)](https://developer.apple.com/documentation/corebluetooth/cbmutableservice/1434330-init)
 
 The following inputs will cause this method to throw a `NSInternalInconsistencyException`:
 * A valid CBUUID with a `data.length` that is not 2 or 16
     * Specifically, a valid CBUUID with `string` length of 8 may be instantiated, but will cause this constructor to throw the error
         * e.g. `CBUUID(string: "ABCD1234")`
-
 
 ## Contributing
